Inherit Records::Asset off Records::Record
A new Layout class is going to be added which will share a lot of code
with Asset. In order to share this code, a new `Record` base class is
required which both classes will inherit from.

The first step is to create the inheritance between Asset and Record
even if it is a blank class.
